+++ history_system/history_core.py
@@ -1,0 +1,3 @@
+from history_manager import HistoryManager
+
+__all__ = ["HistoryManager"]
