+++ history_system/history_descriptions.py
@@ -1,0 +1,11 @@
+from history_manager import HistoryManager
+
+get_anomaly_description = HistoryManager.get_anomaly_description
+get_artifact_base64 = HistoryManager.get_artifact_base64
+_format_timestamp = HistoryManager._format_timestamp
+
+__all__ = [
+    "get_anomaly_description",
+    "get_artifact_base64",
+    "_format_timestamp",
+]
