+++ history_system/history_export.py
@@ -1,0 +1,5 @@
+from history_manager import HistoryManager
+
+export_analysis = HistoryManager.export_analysis
+
+__all__ = ["export_analysis"]
