+++ vifa_pro/forensik_video/config.py
@@ -1,0 +1,3 @@
+from ForensikVideo import CONFIG, Icons
+
+__all__ = ["CONFIG", "Icons"]
