+++ vifa_pro/forensik_video/data_structures.py
@@ -1,0 +1,3 @@
+from ForensikVideo import Evidence, FrameInfo, AnalysisResult
+
+__all__ = ["Evidence", "FrameInfo", "AnalysisResult"]
