+++ vifa_pro/streamlit_ui/components/analysis_tabs.py
@@ -1,0 +1,3 @@
+"""Tab display logic from the original Streamlit app."""
+
+from streamlit_app import *  # noqa: F401,F403
