+++ vifa_pro/streamlit_ui/components/ferm_display.py
@@ -1,0 +1,3 @@
+from streamlit_app import display_ferm_tab_content
+
+__all__ = ["display_ferm_tab_content"]
