+++ vifa_pro/streamlit_ui/components/history_cards.py
@@ -1,0 +1,3 @@
+from streamlit_app import display_history_card
+
+__all__ = ["display_history_card"]
