+++ vifa_pro/streamlit_ui/components/sidebar.py
@@ -1,0 +1,3 @@
+"""Sidebar controls from the original Streamlit app."""
+
+from streamlit_app import *  # noqa: F401,F403
