+++ vifa_pro/streamlit_ui/pages/history_page.py
@@ -1,0 +1,3 @@
+from streamlit_app import render_history_page, display_history_detail
+
+__all__ = ["render_history_page", "display_history_detail"]
