+++ vifa_pro/streamlit_ui/pages/new_analysis.py
@@ -1,0 +1,3 @@
+"""New analysis page logic from the original Streamlit app."""
+
+from streamlit_app import *  # noqa: F401,F403
