+++ vifa_pro/streamlit_ui/styles.py
@@ -1,0 +1,3 @@
+"""CSS styles extracted from the original Streamlit app."""
+
+from streamlit_app import *  # noqa: F401,F403
