+++ vifa_pro/streamlit_ui/utils/metric_descriptions.py
@@ -1,0 +1,3 @@
+from streamlit_app import _get_metric_description
+
+__all__ = ["_get_metric_description"]
